text_embedder_service/requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file='.\requirements.txt' '.\requirements.in'
#
accelerate==1.2.1
    # via -r .\requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via starlette
certifi==2024.12.14
    # via requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via uvicorn
colorama==0.4.6
    # via
    #   click
    #   tqdm
fastapi==0.115.6
    # via -r .\requirements.in
filelock==3.16.1
    # via
    #   huggingface-hub
    #   torch
    #   transformers
fsspec==2024.12.0
    # via
    #   huggingface-hub
    #   torch
h11==0.14.0
    # via uvicorn
huggingface-hub==0.27.1
    # via
    #   accelerate
    #   sentence-transformers
    #   tokenizers
    #   transformers
idna==3.10
    # via
    #   anyio
    #   requests
jinja2==3.1.5
    # via torch
joblib==1.4.2
    # via scikit-learn
markupsafe==3.0.2
    # via jinja2
mpmath==1.3.0
    # via sympy
networkx==3.4.2
    # via torch
numpy==2.2.0
    # via
    #   -r .\requirements.in
    #   accelerate
    #   scikit-learn
    #   scipy
    #   transformers
packaging==24.2
    # via
    #   accelerate
    #   huggingface-hub
    #   transformers
pillow==11.1.0
    # via sentence-transformers
psutil==6.1.1
    # via accelerate
pydantic==2.10.5
    # via fastapi
pydantic-core==2.27.2
    # via pydantic
python-dotenv==1.0.1
    # via -r .\requirements.in
pyyaml==6.0.2
    # via
    #   accelerate
    #   huggingface-hub
    #   transformers
regex==2024.11.6
    # via transformers
requests==2.32.3
    # via
    #   huggingface-hub
    #   transformers
safetensors==0.5.2
    # via
    #   accelerate
    #   transformers
scikit-learn==1.6.1
    # via sentence-transformers
scipy==1.15.1
    # via
    #   scikit-learn
    #   sentence-transformers
sentence-transformers==3.3.1
    # via -r .\requirements.in
sniffio==1.3.1
    # via anyio
starlette==0.41.3
    # via fastapi
sympy==1.13.1
    # via torch
threadpoolctl==3.5.0
    # via scikit-learn
tokenizers==0.21.0
    # via transformers
torch==2.5.1
    # via
    #   -r .\requirements.in
    #   accelerate
    #   sentence-transformers
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   sentence-transformers
    #   transformers
transformers==4.48.0
    # via sentence-transformers
typing-extensions==4.12.2
    # via
    #   anyio
    #   fastapi
    #   huggingface-hub
    #   pydantic
    #   pydantic-core
    #   torch
urllib3==2.3.0
    # via requests
uvicorn==0.34.0
    # via -r .\requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
